requirements.txt
google-cloud-aiplatform>=1.60.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.130.0
python-dotenv>=1.0.0
requests>=2.31.0
